Hi guys,
The working shell "custom_sample_server_app.c" and directory check functions are added. I also added a sample_client_app.c that tests the "fsMount" function as is, I wasn't able to get the subdirectory check working but I have the functions there to do the "fsOpenDir" and "fsReadDir".